app/src/main/java/com/muhammadfarag/popularmovies/CustomArrayAdapter.java
package com.muhammadfarag.popularmovies;

import android.content.Context;
import android.view.LayoutInflater;
import android.view.View;
import android.view.ViewGroup;
import android.widget.BaseAdapter;
import android.widget.ImageView;

import java.util.List;

/**
 * Project: Popular Movies
 * Created by muhammad on 01/07/15.
 */
class CustomArrayAdapter extends BaseAdapter {

    private final Context context;
    private final int resource;
    private final List<String> elements;

    public CustomArrayAdapter(Context context, int resource, List<String> elements) {
        this.context = context;
        this.resource = resource;
        this.elements = elements;
    }

    @Override
    public int getCount() {
        return elements.size();
    }

    @Override
    public Object getItem(int position) {
        return elements.get(position);
    }

    @Override
    public long getItemId(int position) {
        return position;
    }

    @Override
    public View getView(int position, View convertView, ViewGroup parent) {
        ImageView view = (ImageView) convertView;
        if (view == null) {
            view = (ImageView) LayoutInflater.from(this.context).inflate(this.resource, parent, false);
        }

//            String url = getItem(position);

        view.setImageResource(R.drawable.abc_btn_check_to_on_mtrl_000);
        return view;
    }
}
